% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scphaser.R
\name{subset_feat}
\alias{subset_feat}
\title{Subset features}
\usage{
subset_feat(acset, feat)
}
\arguments{
\item{acset}{An acset list created by the \code{\link{new_acset}} function.}

\item{feat}{A character vector with features to subset.}
}
\value{
acset An acset list where every element containing feature-related
information has been subsetted
}
\description{
\code{subset_feat} subsets the features of an acset
}
\examples{
##load dataset
invisible(marinov)
acset = new_acset(featdata = marinov[['featdata']], refcount =
marinov[['refcount']], altcount = marinov[['altcount']], phenodata =
marinov[['phenodata']])

##subset features
sel_feat = c('HMGB1', 'ITGA4')
acset_filt = subset_feat(acset, sel_feat)

}

